Get resumption logic tests into its own module
